Mission: fix float to double conversion when setting lat/lon to NAN

+++ src/modules/navigator/mission.cpp
@@ -1461,8 +1461,8 @@ Mission::do_abort_landing()
 	// XXX: this is a hack to invalidate the "next" position setpoint for the fixed-wing position controller during
 	// the landing abort hold. otherwise, the "next" setpoint would still register as a "LAND" point, and trigger
 	// the early landing configuration (flaps and landing airspeed) during the hold.
-	_navigator->get_position_setpoint_triplet()->next.lat = NAN;
-	_navigator->get_position_setpoint_triplet()->next.lon = NAN;
+	_navigator->get_position_setpoint_triplet()->next.lat = (double)NAN;
+	_navigator->get_position_setpoint_triplet()->next.lon = (double)NAN;
 	_navigator->get_position_setpoint_triplet()->next.alt = NAN;
 
 	publish_navigator_mission_item(); // for logging
